Create ThreadPool

+++ PlayerServer/Socket.h
@@ -104,7 +104,8 @@ public:
 	virtual int Close() {
 		m_status = 3;
 		if (m_socket != -1) {
-			unlink(m_param.ip);
+			if (m_param.attr & SOCK_ISSERVER)
+				unlink(m_param.ip);
 			int fd = m_socket;
 			m_socket = -1;
 			close(fd);
@@ -210,7 +211,7 @@ public:
 			return (int)len;//收到数据
 		}
 		if (len < 0) {
-			if (errno == EINTR || (errno == EAGAIN)) {
+			if (errno == EINTR || (errno == EAGAIN)) {//非阻塞
 				data.clear();
 				return 0;//没有数据收到
 			}
